Added backend to frontend communication
Made changes in the frontend so that the dashboard can communicate with the backend const partners to map out each PartnerTile and display that. Also added 'active' as a key to the partners const/JSON so as to keep track of whether C4C is actively working with that partner. Updated CSS to contain  true and false colors.